[REFACTOR]: Repository에서 Mapper로 연결
