+++ pickme/src/main/java/project/pickme/item/mapper/LawMapper.java
@@ -1,4 +1,4 @@
-package project.pickme.item.repository;
+package project.pickme.item.mapper;
 
 import java.util.List;
