+++ pickme/src/main/java/project/pickme/item/service/LawService.java
@@ -2,12 +2,11 @@ package project.pickme.item.service;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import project.pickme.item.dto.LawDto;
-import project.pickme.item.repository.LawMapper;
+import project.pickme.item.mapper.LawMapper;
 
 @Service
 @RequiredArgsConstructor
